Implement `declare major` and stubs for `module` commands
`declare major` has been implemented together with serialization of the
major.

Stubs for `module add`, `module remove` and `module list` have been
added. The module codes are syntatically validated to be alphanumeric, but are not
yet validated for existence in module database.

Serialization is changed slightly with addition of `activeStudent`
field.

Some reorganization of code is done, but this is restricted to reduce merge
conflicts with other PRs.