Add `timetable add, remove`, checks to `module`
Timetables can be added and removed with the
`timetable add/remove year/[YEAR] sem/[SEM]` sub-commands respectively.
Validation has also been added to `module` sub-commands to reject the
command if no student or timetable is active.

A known bug is that the timetables of a `Student` instance isn't saved
in the `students` list in JSON. This is due to `UniqueStudentList#get`
(implemented by `ObservableList<Student>`) returning a copy of the `Student` retrieved.
This will be fixed in a future commit.