Reorder structure of UG
This is due to a suggestion from the PE-D https://github.com/AY1920S2-CS2103T-W13-4/main/issues/174 .
Now Student Management is before Degree and Grade Management, which is
before Timetable and then Module management.

This lets the user read the UG from top to bottom (with the exception of
`module grade`, which has a link to the Module Management section).